Refactor UTS field with FixedCStr
