+++ kernel/src/syscall/sethostname.rs
@@ -1,9 +1,12 @@
 // SPDX-License-Identifier: MPL-2.0
 
-use crate::{net::uts_ns::UtsField, prelude::*, syscall::SyscallReturn};
+use crate::{
+    prelude::*,
+    syscall::{SyscallReturn, setdomainname::read_uts_field},
+};
 
 pub fn sys_sethostname(addr: Vaddr, len: usize, ctx: &Context) -> Result<SyscallReturn> {
-    let new_host_name = UtsField::read_from(addr, len, ctx)?;
+    let new_host_name = read_uts_field(addr, len, ctx)?;
 
     let ns_proxy_ref = ctx.thread_local.borrow_ns_proxy();
     let ns_proxy = ns_proxy_ref.unwrap();
